pyproject.toml
[project]
name = "boardroom"
version = "0.1.0"
description = "Multi-agent financial analysis system"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "websockets>=14.0",
    "langgraph>=0.2.0",
    "anthropic>=0.40.0",
    "openai>=1.55.0",
    "google-genai>=1.0.0",
    "yfinance>=0.2.50",
    "sqlalchemy[asyncio]>=2.0.0",
    "asyncpg>=0.30.0",
    "alembic>=1.14.0",
    "pydantic-settings>=2.6.0",
    "httpx>=0.28.0",
    "numpy>=2.0.0",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.0",
    "bcrypt>=4.0.0,<5.0.0",
    "loguru>=0.7.0",
    "python-dotenv>=1.0.0",
    "python-multipart>=0.0.9",
    "pytest>=9.0.2",
    "email-validator>=2.3.0",
    "redis>=7.1.0",
    "litellm>=1.81.13",
    "pyyaml>=6.0.3",
    "langfuse>=3.14.2",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=6.0.0",
    "ruff>=0.2.0",
    "mypy>=1.8.0",
    "black>=24.0.0",
]

[dependency-groups]
dev = [
    "aiosqlite>=0.22.1",
    "pytest-asyncio>=1.3.0",
    "ruff>=0.2.0",
    "mypy>=1.8.0",
    "black>=24.0.0",
]

[tool.black]
line-length = 88
target-version = ['py313']
include = '\.pyi?$'
exclude = '''
/(
    \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
line_length = 88
multi_line_mode = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true
skip_glob = ["*/migrations/*", ".venv/*"]

[tool.mypy]
python_version = "3.13"
check_untyped_defs = true
ignore_missing_imports = true
warn_unused_ignores = true
warn_redundant_casts = true
warn_unused_configs = true
disallow_incomplete_defs = false
disallow_untyped_defs = false

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
asyncio_default_fixture_scope = "function"
markers = [
    "unit: Unit tests (fast, no external deps)",
    "integration: Integration tests (requires PostgreSQL)",
]
addopts = "--strict-markers -ra"

[tool.coverage.run]
source = ["backend"]
omit = [
    "*/tests/*",
    "*/test_*.py",
    "*/__init__.py",
    "*/migrations/*",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
    "@abstractmethod",
    "pass",
    "\\.\\.\\.",
]
fail_under = 80
precision = 2
